fix(MDS064): rename package headingwhitespace → atxheadingwhitespace; register in all.go
The package name headingwhitespace did not follow the repo convention
of deriving the directory name from the rule name with hyphens removed
(blockquote-whitespace → blockquotewhitespace, list-marker-space →
listmarkerspace). Rename to atxheadingwhitespace for consistency.

Also fixes a production bug: the package was never blank-imported in
internal/rules/all/all.go, so MDS064 was not registered and would
never run in cmd/mdsmith. Add the import in alphabetical order.

Also updates plan/176 front matter and design text to replace the
provisional MDS060 ID with the shipped MDS064.
