+++ internal/integration/rules_test.go
@@ -38,7 +38,7 @@ import (
 	_ "github.com/jeduden/mdsmith/internal/rules/githooksync"
 	_ "github.com/jeduden/mdsmith/internal/rules/headingincrement"
 	_ "github.com/jeduden/mdsmith/internal/rules/headingstyle"
-	_ "github.com/jeduden/mdsmith/internal/rules/headingwhitespace"
+	_ "github.com/jeduden/mdsmith/internal/rules/atxheadingwhitespace"
 	_ "github.com/jeduden/mdsmith/internal/rules/horizontalrulestyle"
 	_ "github.com/jeduden/mdsmith/internal/rules/include"
 	_ "github.com/jeduden/mdsmith/internal/rules/linelength"
