+++ internal/rules/atxheadingwhitespace/rule_test.go
@@ -1,4 +1,4 @@
-package headingwhitespace
+package atxheadingwhitespace
 
 import (
 	"testing"
